requirements.txt
absl-py==1.0.0
astor==0.8.1
bleach==1.5.0
cached-property==1.5.2
click==8.0.4
dataclasses==0.8
docopt==0.6.2
Flask==2.0.3
Flask-SQLAlchemy==2.5.1
gast==0.2.2
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.44.0
h5py==3.1.0
html5lib==0.9999999
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.3.6
MarkupSafe==2.0.1
mock==4.0.3
numpy==1.16.4
opt-einsum==3.3.0
pandas==1.1.5
protobuf==3.19.4
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2022.1
scikit-learn==0.24.2
scipy==1.5.4
six==1.16.0
SQLAlchemy==1.4.38
tensorboard==1.13.1
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.1
termcolor==1.1.0
threadpoolctl==3.1.0
typing_extensions==4.1.1
Werkzeug==2.0.3
wrapt==1.14.0
zipp==3.6.0
